Remove old/unused "debug" section from config.txt
The "log_maxlen" value of the "debug" section was used to tune the
maximum amount of log messages that would be buffered into the log
deque. With https://github.com/gridsync/gridsync/pull/570, however,
logs are persisted to disk (rather than buffered into memory) and
the "[debug]log_maxlen" value is ignored.